Extract Analytics types to a separate file
